Removed TEST from Debug, Added KSP dlls to ignore
I removed TEST from the "Debug" Profile so now it will actually respond
to #define TEST being commented out. this means Debug can now build a
KSP compatible DLL and also a DLL that can be tested (using testGHud)
without KSP. There will be a little bit more to do with this in the
readme under the link in the Dev Tips heading in the next commit
shortly.
I also added a statement to the .gitignore to ignore any KSP Dll's in
the KSPFiles folder.
Most of the commits for now will be general house cleaning then I expect
to be doing some code overhauls