Instantiates once, More Feedback, Higher FPS
Didn't include the fix (hopefully) for crashes keeping the LCD applet
open until a powercycle of the keyboard or computer in this commit, will
commit it either here or in another branch. Also increased the framerate
so it is less jerky (AA comes next).
After the crash fix is commited, I will work on a code overhaul (The
general idea for most of the code is there but just needs to be
changed/added to).