Crash Fix (Hopefully Works!)
How this is supposed to work is that when the plugin starts up, it
stores the LCD connection data in a file and deletes it when it closes
down. If KSP crashes then that data should still be in the file and it
will try and close the connections. It is untested as I can't find a way
to crash KSP, but theoretically it should work.